refactor(desktop): extract config.rs from commands.rs (#235)
Move the .tiki/config.json read/save logic — 7 config structs, the
unknown-key warning collector, the read_tiki_config/save_tiki_config
commands, and their 6 tests — out of the 965-line commands.rs into a
dedicated config.rs module. The config types had zero references outside
commands.rs, so this is a pure cohesion split with no behavior change.

- new apps/desktop/src-tauri/src/config.rs (structs + commands + tests)
- commands.rs: 965 -> 686 lines; resolve_tiki_path is now pub(crate) so
  config.rs can reuse it
- lib.rs: declare `mod config;` and register config::{read,save}_tiki_config

cargo clippy --deny warnings clean; cargo test green (config_tests 6/6).

Closes #235
